+++ ClientApplicationMVC/Views/Home/Index.cshtml
@@ -8,7 +8,7 @@
 
 <div class="w3-container w3-display-middle">
     <h1 class="w3-animate-fading">
-        <a style=color:#785ff4>SENG </a><a style=color:#1185f9>401 </a><a style=color:#3ae2fc>Project</a> Login Page
+        <a style=color:#785ff4>SENG </a><a style=color:#1185f9>401 </a><a style=color:#3ae2fc>Project</a> Home Page
     </h1>
     <p>
         Call It Local is a project created by the students
