+++ Messages/DataTypes/Database/CompanyDirectory/CompanyInstance.cs
@@ -1,5 +1,4 @@
 ﻿using NServiceBus;
-
 using System;
 
 namespace Messages.DataTypes.Database.CompanyDirectory
