+++ Messages/DataTypes/Database/Weather/Imperial.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Messages.DataTypes.Database.Weather
+{
+    [Serializable]
+    public class Imperial
+    {
+        public float Value { get; set; }
+        public string Unit { get; set; }
+        public int UnitType { get; set; }
+    }
+}
