+++ Messages/DataTypes/Database/Weather/Temperature.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Messages.DataTypes.Database.Weather
+{
+    [Serializable]
+    public class Temperature
+    {
+        public Metric Metric { get; set; }
+        public Imperial Imperial { get; set; }
+    }
+}
