+++ Messages/DataTypes/Database/Weather/WeatherInfo.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+
+namespace Messages.DataTypes.Database.Weather
+{
+    [Serializable]
+    public class WeatherInfo
+    {
+        public DateTime LocalObservationDateTime { get; set; }
+        public int EpochTime { get; set; }
+        public string WeatherText { get; set; }
+        public int WeatherIcon { get; set; }
+        public bool IsDayTime { get; set; }
+        public Temperature Temperature { get; set; }
+        public string MobileLink { get; set; }
+        public string Link { get; set; }
+    }
+}
